# docs/swagger.yaml
basePath: /swagger/box-manage-service
definitions:
  controllers.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: 详细错误信息
        type: string
      message:
        example: 参数错误
        type: string
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
    type: object
...
  controllers.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      timestamp:
        example: "2025-01-26T12:00:00Z"
        type: string
      total:
        example: 100
        type: integer
    type: object
...
  models.ConvertedModel:
    description: 转换后模型数据结构，支持bmodel等转换后格式的完整管理
    properties:
      access_count:
        description: 访问控制和统计
        example: 50
        type: integer
      conversion_task_id:
        description: 关联转换任务ID
        example: conv_task_abc123
        type: string
      convert_params:
        description: 转换参数
        example: '{}'
        type: string
      converted_at:
        description: 转换完成时间
        example: "2025-01-28T12:00:00Z"
        type: string
      converted_path:
        description: 转换后文件路径（与FilePath相同，为了兼容数据库）
        example: /data/models/converted/yolo-v5-bm1684x.bmodel
        type: string
      created_at:
        description: 创建时间
        example: "2025-01-26T12:00:00Z"
        type: string
      description:
        description: 模型描述
        example: 转换为BM1684X芯片优化的YOLO v5模型
        type: string
      display_name:
        description: 显示名称
        example: YOLO v5 目标检测模型 (BM1684X)
        type: string
      download_count:
        description: 下载次数
        example: 10
        type: integer
      file_md5:
        description: 文件MD5校验和
        example: d41d8cd98f00b204e9800998ecf8427e
        type: string
      file_name:
        description: 文件信息
        example: yolo-v5-bm1684x.bmodel
        type: string
      file_path:
        description: 文件存储路径
        example: /data/models/converted/yolo-v5-bm1684x.bmodel
        type: string
      file_sha256:
        description: 文件SHA256校验和
        example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        type: string
      file_size:
        description: 文件大小（字节）
        example: 52428800
        type: integer
      id:
        description: 主键ID
        example: 1
        type: integer
      input_channels:
        description: 输入通道数
        example: 3
        type: integer
      input_height:
        description: 输入高度
        example: 640
        type: integer
      input_width:
        description: 输入宽度
        example: 640
        type: integer
      last_accessed_at:
        description: 最后访问时间
        example: "2025-01-28T12:00:00Z"
        type: string
      model_key:
        description: '模型标识 - 格式: type-targetyoloVersion-targetChip-name'
        example: detection-yolov8-bm1684x-cow_head
        type: string
      name:
        description: 基本信息
        example: yolo-v5-bm1684x
        type: string
      original_model:
        allOf:
        - $ref: '#/definitions/models.OriginalModel'
        description: 关联模型
      original_model_id:
        description: 关联信息
        example: 1
        type: integer
      quantization:
        description: 量化类型 (F16/F32)
        example: F16
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ConvertedModelStatus'
        description: 状态管理
        example: completed
      tags:
        description: 元数据和标签
        example: optimized,production,bm1684x
        type: string
      target_chip:
        description: 目标芯片
        example: bm1684x
        type: string
      target_yolo_version:
        description: 目标YOLO版本
        example: yolov8
        type: string
      task_type:
        allOf:
        - $ref: '#/definitions/models.ModelTaskType'
        description: AI任务属性（继承自原始模型）
        example: detection
      updated_at:
        description: 更新时间
        example: "2025-01-26T12:00:00Z"
        type: string
      user_id:
        description: 创建用户ID
        example: 1
        type: integer
      version:
        description: 版本号
        example: 1.0.0
        type: string
    type: object
  models.ConvertedModelStatus:
    enum:
    - pending
    - converting
    - completed
    - failed
    type: string
    x-enum-comments:
      ConvertedModelStatusCompleted: 已完成
      ConvertedModelStatusConverting: 转换中
      ConvertedModelStatusFailed: 转换失败
      ConvertedModelStatusPending: 待转换
    x-enum-descriptions:
    - 待转换
    - 转换中
    - 已完成
    - 转换失败
    x-enum-varnames:
    - ConvertedModelStatusPending
    - ConvertedModelStatusConverting
    - ConvertedModelStatusCompleted
    - ConvertedModelStatusFailed
...
  models.ModelBoxDeployment:
    description: 记录转换后模型在哪些盒子上已成功部署
    properties:
      box:
        $ref: '#/definitions/models.Box'
      box_id:
        description: 盒子ID
        type: integer
      converted_model:
        allOf:
        - $ref: '#/definitions/models.ConvertedModel'
        description: 关联对象
      converted_model_id:
        description: 关联信息
        type: integer
      created_at:
        description: 创建时间
        example: "2025-01-26T12:00:00Z"
        type: string
      deployed_at:
        description: 部署信息
        type: string
      deployment_path:
        description: 部署路径
        type: string
      id:
        description: 主键ID
        example: 1
        type: integer
      is_verified:
        description: 验证信息
        type: boolean
      last_verified_at:
        description: 最后验证时间
        type: string
      model_key:
        description: 模型Key
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ModelBoxDeploymentStatus'
        description: 部署状态
      task_id:
        description: 部署任务ID
        type: string
      updated_at:
        description: 更新时间
        example: "2025-01-26T12:00:00Z"
        type: string
      verification_msg:
        description: 验证消息
        type: string
      version:
        description: 部署时的模型版本
        type: string
    type: object
...
  service.ConvertedModelStatistics:
    properties:
      active_models:
        type: integer
      chip_stats:
        additionalProperties:
          format: int64
          type: integer
        type: object
      deployed_models:
        type: integer
      status_stats:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total_models:
        type: integer
    type: object
...
  service.CreateConvertedModelRequest:
    properties:
      conversion_task_id:
        type: string
      convert_params:
        type: string
      description:
        type: string
      display_name:
        type: string
      file_md5:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_sha256:
        type: string
      file_size:
        type: integer
      name:
        type: string
      original_model_id:
        type: integer
      quantization:
        type: string
      tags:
        items:
          type: string
        type: array
      user_id:
        type: integer
      version:
        type: string
    required:
    - file_name
    - file_path
    - name
    - original_model_id
    - user_id
    - version
    type: object
...
  service.UpdateConvertedModelRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: AI盒子管理系统后端服务，提供盒子管理、模型管理、任务管理、用户管理等功能
  title: AI盒子管理系统 API
  version: "1.0"
paths:
...
  /api/converted-models:
    get:
      consumes:
      - application/json
      description: 分页获取转换后模型列表，支持多条件搜索
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 模型状态
        enum:
        - active
        - inactive
        - failed
        in: query
        name: status
        type: string
      - description: 目标芯片
        enum:
        - bm1684
        - bm1684x
        - bm1688
        in: query
        name: target_chip
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 原始模型ID
        in: query
        name: original_model_id
        type: integer
      - description: 标签（逗号分隔）
        in: query
        name: tags
        type: string
      - description: 模型名称搜索关键词
        in: query
        name: name
        type: string
      - description: 版本搜索关键词
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ConvertedModel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 获取转换后模型列表
      tags:
      - 转换后模型
    post:
      consumes:
      - application/json
      description: 创建新的转换后模型记录
      parameters:
      - description: 转换后模型信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateConvertedModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ConvertedModel'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 创建转换后模型
      tags:
      - 转换后模型
  /api/converted-models/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的转换后模型
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 删除转换后模型
      tags:
      - 转换后模型
    get:
      consumes:
      - application/json
      description: 根据ID获取转换后模型详情
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ConvertedModel'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 获取转换后模型详情
      tags:
      - 转换后模型
    put:
      consumes:
      - application/json
      description: 更新转换后模型信息
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateConvertedModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 更新转换后模型
      tags:
      - 转换后模型
  /api/converted-models/{id}/deployments:
    get:
      consumes:
      - application/json
      description: 根据转换后模型ID获取其在哪些盒子上已部署
      parameters:
      - description: 转换后模型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ModelBoxDeployment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 获取转换后模型的部署信息
      tags:
      - 转换后模型
  /api/converted-models/{id}/download:
    get:
      consumes:
      - application/json
      description: 下载转换后模型文件
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 模型文件
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 下载转换后模型
      tags:
      - 转换后模型
  /api/converted-models/key/{modelKey}/deployments:
    get:
      consumes:
      - application/json
      description: 根据模型Key获取其在哪些盒子上已部署
      parameters:
      - description: 模型Key
        in: path
        name: modelKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ModelBoxDeployment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 根据模型Key获取部署信息
      tags:
      - 转换后模型
  /api/converted-models/original/{originalModelId}:
    get:
      consumes:
      - application/json
      description: 根据原始模型ID获取相关的转换后模型列表
      parameters:
      - description: 原始模型ID
        in: path
        name: originalModelId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ConvertedModel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 根据原始模型获取转换后模型
      tags:
      - 转换后模型
  /api/converted-models/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索转换后模型
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - description: 目标芯片
        in: query
        name: target_chip
        type: string
      - description: 模型状态
        in: query
        name: status
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ConvertedModel'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 搜索转换后模型
      tags:
      - 转换后模型
  /api/converted-models/statistics:
    get:
      consumes:
      - application/json
      description: 获取转换后模型的统计信息
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.ConvertedModelStatistics'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 获取转换后模型统计
      tags:
      - 转换后模型
...
swagger: "2.0"